Tests/LibWeb/Text/expected/indexeddb-upgrade-multiple-connections-close-ordering.txt

setup.upgradeneeded
setup.success
second.success
connA.versionchange
connA.close() called
connB.versionchange
connB.close() called
upgrade.upgradeneeded
upgrade.success
DONE
